+++ CMakeLists.txt
@@ -4,14 +4,17 @@ project(Pacman LANGUAGES C)
 
 find_package(SDL2 REQUIRED)
 
-add_executable(${PROJECT_NAME} "" pacman.h)
-target_sources(${PROJECT_NAME} PRIVATE pacman.c)
+add_executable(${PROJECT_NAME} "" src/pacman.h src/main.h src/menu_window.h src/pacman_window.h src/sdl_functions.h)
+target_sources(${PROJECT_NAME} PRIVATE src/main.c src/menu_window.c src/pacman_window.c src/main.c src/sdl_functions.c)
 target_include_directories(${PROJECT_NAME} PRIVATE ${SDL2_INCLUDE_DIRS})
 target_compile_definitions(${PROJECT_NAME} PRIVATE "SDL_MAIN_HANDLED")
 target_link_libraries(${PROJECT_NAME} PRIVATE ${SDL2_LIBRARIES})
 
-file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/pacman_sprites.bmp
-	DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/)	
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/assets/img/pacman_sprites.bmp
+	DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/)
+
+file(COPY ${CMAKE_CURRENT_SOURCE_DIR}/assets/img/play_btn.bmp
+		DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/)
 
 IF (MSVC)
 	set_property(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR} PROPERTY VS_STARTUP_PROJECT ${PROJECT_NAME})
