+++ src/main.c
@@ -1,0 +1,14 @@
+#include "main.h"
+
+int main(__attribute__((unused)) int argc, __attribute__((unused)) char** argv)
+{
+    initSDL();
+
+    loadMedia();
+
+    displayMainMenu();
+
+    closeWindow();
+
+    return 0;
+}
