+++ src/main.h
@@ -1,0 +1,10 @@
+//
+// Created by Julien on 09/02/2023.
+//
+
+#ifndef PACMAN_MAIN_H
+#define PACMAN_MAIN_H
+
+#include "pacman.h"
+
+#endif //PACMAN_MAIN_H
