+++ src/pacman_window.h
@@ -1,0 +1,14 @@
+//
+// Created by Julien on 09/02/2023.
+//
+
+#ifndef PACMAN_PACMAN_WINDOW_H
+#define PACMAN_PACMAN_WINDOW_H
+
+#include "pacman.h"
+
+void startGameLoop();
+bool handleGameEvents();
+void draw();
+
+#endif //PACMAN_PACMAN_WINDOW_H
