+++ src/sdl_functions.h
@@ -1,0 +1,15 @@
+//
+// Created by Julien on 09/02/2023.
+//
+
+#ifndef PACMAN_SDL_FUNCTIONS_H
+#define PACMAN_SDL_FUNCTIONS_H
+
+#include "pacman.h"
+
+void initSDL();
+void loadMedia();
+void SDL_ExitWithError(const char *message);
+void closeWindow();
+
+#endif //PACMAN_SDL_FUNCTIONS_H
